Update damn old ACE (lite version) to 6.3.3
This update fixes issues in ACE before 6.3.3 version.
ACE continues to build only under windows but probably
on sonn will on other platforms too - needs some work,
since not all platform needs all files.

ACE is now compilable with VS 2015

+++ dep/acelite/ace/Configuration_Import_Export.cpp
@@ -1,5 +1,3 @@
-// $Id: Configuration_Import_Export.cpp 84565 2009-02-23 08:20:39Z johnnyw $
-
 #include "ace/Configuration_Import_Export.h"
 #include "ace/OS_Errno.h"
 #include "ace/OS_NS_stdio.h"
@@ -138,7 +136,8 @@ ACE_Registry_ImpExp::import_config (const ACE_TCHAR* filename)
               // number type
               ACE_TCHAR* endptr = 0;
               unsigned long value = ACE_OS::strtoul (end + 6, &endptr, 16);
-              if (config_.set_integer_value (section, name, value))
+              if (config_.set_integer_value (section, name,
+                                             static_cast<u_int> (value)))
                 {
                   ACE_OS::fclose (in);
                   delete [] buffer;
